started refactoring code into hodlcore and implementing cron jobs
added database with some prefilled data

more refactoring

json formatting
