+++ README.md
@@ -1,2 +1,4 @@
 # hodlmybot
 Telegram bot with a wide variety of features related to crypto currency.
+
+Requred dependencies: tweepy, bs4
