--- data/README.md
@@ -1,1 +1,0 @@
-# Data dir
