+++ defaults-reddit.json
@@ -1,0 +1,26 @@
+[
+	"bitcoin",
+	"ethereum",
+	"litecoin",
+	"NEO",
+	"nebulas",
+	"Monero",
+	"Ripple",
+	"IOTA",
+	"nanocurrency",
+	"vertcoin",
+	"HEROcoin",
+	"XRPTalk",
+	"Electroneum",
+	"dogecoin",
+	"reddCoin",
+	"Lisk",
+	"siacoin",
+	"steem",
+	"komodoplatform",
+	"SaltCoin",
+	"gnosisPM",
+	"Quantstamp",
+	"COSS",
+	"CossIO"
+]
