+++ defaults-twitter.json
@@ -1,0 +1,293 @@
+[
+	"ethereumproject",
+	"Ripple",
+	"LitecoinProject",
+	"CardanoStiftung",
+	"neo_blockchain",
+	"StellarOrg",
+	"eos_io",
+	"Dashpay",
+	"iotatoken",
+	"monerocurrency",
+	"NEMofficial",
+	"eth_classic",
+	"tronfoundation",
+	"vechainofficial",
+	"LiskHQ",
+	"tether_to",
+	"bitcoingold",
+	"qtumOfficial",
+	"omise_go",
+	"helloiconworld",
+	"zcashco",
+	"nanocurrency",
+	"vergecurrency",
+	"binance_2017",
+	"Bytecoin_BCN",
+	"Steemit",
+	"BitPopulous",
+	"SiaTechHQ",
+	"stratisplatform",
+	"dogecoin",
+	"wavesplatform",
+	"rchain_coop",
+	"ethstatus",
+	"digixglobal",
+	"MakerDAO",
+	"BitShares",
+	"aetrnty",
+	"Veritaseuminc",
+	"Waltonchain",
+	"AugurProject",
+	"decredproject",
+	"ZclassicCoin",
+	"electroneum",
+	"0xproject",
+	"HcashOfficial",
+	"revain_org",
+	"ardorplatform",
+	"KomodoPlatform",
+	"ArkEcosystem",
+	"kucoincom",
+	"attentiontoken",
+	"DigiByteCoin",
+	"zilliqa",
+	"Cryptonex_CNX",
+	"loopringorg",
+	"Bytom_Official",
+	"aelfblockchain",
+	"dragonchaingang",
+	"dentcoin",
+	"syscoin",
+	"dentacoin",
+	"kybernetwork",
+	"polymathnetwork",
+	"quoine_SG",
+	"nebulasio",
+	"golemproject",
+	"_pivx",
+	"Aion_Network",
+	"ByteballOrg",
+	"ethos_io",
+	"iostoken",
+	"reddcoin",
+	"Bitcore_BTX",
+	"gongxinbao",
+	"factom",
+	"powerledger_io",
+	"FunFairTech",
+	"Crowd_indicator",
+	"kin_foundation",
+	"smart_contract",
+	"PillarWallet",
+	"zcoinofficial",
+	"SaltLending",
+	"NxtCommunity",
+	"BancorNetwork",
+	"enigmampc",
+	"maidsafe",
+	"ParticlProject",
+	"NeblioTeam",
+	"scashofficial",
+	"emercoin_press",
+	"requestnetwork",
+	"Vertcoin",
+	"tenxwallet",
+	"ignisguide",
+	"The_Blocknet",
+	"gamecredits",
+	"Quantstamp",
+	"raiden_network",
+	"WAX_io",
+	"singularity_net",
+	"iconominet",
+	"XDNCommunity",
+	"enjin",
+	"gnosisPM",
+	"santimentfeed",
+	"_poetproject",
+	"BitcoinDark",
+	"thebigxp",
+	"Skycoinproject",
+	"SuperNETorg",
+	"civickey",
+	"storjproject",
+	"substratumnet",
+	"CryptoBridge",
+	"achainofficial",
+	"Storm_Token",
+	"ZenCashOfficial",
+	"NxsEarth",
+	"decentraland",
+	"ethlend1",
+	"AragonProject",
+	"HPB_Global",
+	"OysterProtocol",
+	"DewFund",
+	"AdEx_Network",
+	"NAVCoin",
+	"XPAtwopointoh",
+	"telcoin_team",
+	"TokensNet",
+	"_medibloc",
+	"blockv_io",
+	"TimeNewBank",
+	"iEx_ec",
+	"IoT_Chain",
+	"monaco_card",
+	"ubiqsmart",
+	"bluzellehq",
+	"MediShares",
+	"Sophia_TX_",
+	"Blockchain_Data",
+	"PayPiePlatform",
+	"genesis_vision",
+	"red_pulse_china",
+	"Bibox365",
+	"udotcash",
+	"VibeHubVR",
+	"PeercoinPPC",
+	"wagerrx",
+	"SIRINLABS",
+	"AschPlatform",
+	"HTMLCOIN",
+	"AmbrosusAMB",
+	"RCN_token",
+	"sonmdevelopment",
+	"CRYPTOtwenty",
+	"PuraSocial",
+	"DeepBrainChain",
+	"eidoo_io",
+	"xtrabytes",
+	"metalpaysme",
+	"TauChainOrg",
+	"thetatoken",
+	"appcoinsproject",
+	"einsteiniumcoin",
+	"jibrelnetwork",
+	"cybermiles",
+	"airswap",
+	"SingularDTV",
+	"etherparty_com",
+	"melonport",
+	"TheSimpleToken",
+	"wabiico",
+	"QRLedger",
+	"minexcoin",
+	"CounterpartyXCP",
+	"streamrinc",
+	"mvs_org",
+	"ins_ecosystem",
+	"cobinhood",
+	"GIFTO_io",
+	"PaccoinOfficial",
+	"BitBayofficial",
+	"thenagaico",
+	"blocksafe",
+	"edgelessproject",
+	"nulsservice",
+	"TrinityProtocol",
+	"viacoin",
+	"breadapp",
+	"UTRUST_Official",
+	"WePowerN",
+	"mobilegotoken",
+	"Burstcoin_dev",
+	"AeonCoin",
+	"wingsplatform",
+	"Tierion",
+	"coindashio",
+	"bottos_ai",
+	"project_ecc",
+	"gulden",
+	"spankchain",
+	"Centra_Card",
+	"medical_chain",
+	"CloakCoin",
+	"ionomics",
+	"modum_io",
+	"delphy_org",
+	"indaHash",
+	"TaaSfund",
+	"lbryio",
+	"hiveproject_net",
+	"district0x",
+	"BitBeanCoin",
+	"qlinkmobi",
+	"RiseVisionTeam",
+	"GroestlcoinTeam",
+	"allsportschain",
+	"intchain",
+	"ATMChainDev",
+	"CrownPlatform",
+	"safe_exchange",
+	"cosscrypto",
+	"myrarepepe",
+	"ColossusCoinXT",
+	"UnikoinGold",
+	"lykke",
+	"Feathercoin",
+	"cappasity",
+	"mercurytoken",
+	"stktoken",
+	"tokencard_io",
+	"SpectreAI",
+	"dmdcoin",
+	"Namecoin",
+	"ad_chain",
+	"HorizonState",
+	"nimiq",
+	"presearchnews",
+	"origin_trail",
+	"Viberate_com",
+	"datumnetwork",
+	"mooncoinitalia",
+	"Dimecoin_",
+	"DECENTplatform",
+	"potcoin",
+	"SibChervonec",
+	"firstbloodio",
+	"DeepOnionx",
+	"Humaniq",
+	"LunyrInc",
+	"everexio",
+	"Monetha_io",
+	"EnergoOfficial",
+	"blockmasonio",
+	"TheHempCoin ",
+	"Qbao2339",
+	"AgrelloOfficial",
+	"WhiteCoiner",
+	"ODYSSEYPROTOCOL",
+	"ShiftNrg",
+	"inklabsfound",
+	"Blockportio",
+	"cofound_it",
+	"swftcoin",
+	"zeepinchain",
+	"Zeusshield",
+	"io_coin",
+	"mintcointeam",
+	"FlashCoins",
+	"unobtanium_uno",
+	"TheSwarmFund",
+	"SelfKey",
+	"ElectracoinECA",
+	"bitconnect",
+	"soarcoin",
+	"lamdentau",
+	"gelert",
+	"BloomToken",
+	"Xspec_Official",
+	"PascalCoin",
+	"blocktix",
+	"numerai",
+	"SHIELDCurrency",
+	"carVertical_com",
+	"worldcoresocial",
+	"sun_contract",
+	"dadi",
+	"Voxelus",
+	"wetrustplatform",
+	"maecenasart"
+]
